requirements.txt
openai
python-dotenv
requests
python-wordpress-xmlrpc

